把拆红包方法放到红包工具类里
